fix utils split
